Redirect favorites page
[added] Redirect favorites page
[added] Added IndexedStack Widget
