+++ lib/config/router/app_router.dart
@@ -2,12 +2,16 @@ import 'package:cinemapedia/presentation/screens/screens.dart';
 import 'package:go_router/go_router.dart';
 
 final appRouter = GoRouter(
-  initialLocation: '/',
+  initialLocation: '/home/0',
   routes: [
     GoRoute(
-        path: '/',
+        path: '/home/:page',
         name: HomeScreen.name,
-        builder: (context, state) => const HomeScreen(),
+        builder: (context, state) {
+          final pageIndex = state.pathParameters['page'] ?? '0';
+
+          return HomeScreen(pageIndex: int.parse(pageIndex));
+        },
         routes: [
           GoRoute(
             path: 'movie/:id',
@@ -19,5 +23,9 @@ final appRouter = GoRouter(
             },
           ),
         ]),
+    GoRoute(
+      path: '/',
+      redirect: (_, __) => '/home/0',
+    ),
   ],
 );
