+++ lib/presentation/widgets/movies/movies_slideshow.dart
@@ -60,7 +60,7 @@ class _Slide extends StatelessWidget {
 
                   // return FadeIn(child: child);
                   return GestureDetector(
-                    onTap: () => context.push('/movie/${movie.id}'),
+                    onTap: () => context.push('/home/0/movie/${movie.id}'),
                     child: FadeIn(child: child),
                   );
                 },
