+++ lib/presentation/widgets/shared/custom_bottom_navigation.dart
@@ -1,26 +1,27 @@
 import 'package:flutter/material.dart';
+import 'package:go_router/go_router.dart';
 
 class CustomBottomNavigation extends StatelessWidget {
-  const CustomBottomNavigation({super.key});
+  final int currentIndex;
+
+  const CustomBottomNavigation({
+    super.key,
+    required this.currentIndex,
+  });
 
   @override
   Widget build(BuildContext context) {
     return BottomNavigationBar(
       elevation: 0,
+      currentIndex: currentIndex,
+      onTap: (value) => context.go('/home/$value'),
       items: const [
+        BottomNavigationBarItem(icon: Icon(Icons.home_max), label: 'Inicio'),
         BottomNavigationBarItem(
-          icon: Icon(Icons.home_max),
-          label: 'Inicio'
-        ),
+            icon: Icon(Icons.label_outline), label: 'Categorías'),
         BottomNavigationBarItem(
-          icon: Icon(Icons.label_outline),
-          label: 'Categorías'
-        ),
-        BottomNavigationBarItem(
-          icon: Icon(Icons.favorite_outline),
-          label: 'Favoritos'
-        ),
-      ]
+            icon: Icon(Icons.favorite_outline), label: 'Favoritos'),
+      ],
     );
   }
 }
